[#40753] Set a small maximum timeout for db connections
Before the call too `./manage.py ensure_db_connection --wait 30` in
`docker-entrypoint.sh` could take an unspecified amount of time based on the
configuration of the host. One machine was observed having a ~30 second timeout
between of the 30 steps. Another was observed having a ~2 minute timeout between
each steps. That would take the container ~15min and ~1 hour respectively to
crash on startup.

This sets the postgres timeout to the minimum allowed of 2 seconds. With the 1
second `wait` in `ensure_db_connection` it sums to a maximum of 3 second seconds
per step. Giving the container a maximum wait time on startup of about 45
seconds.