Add HTTP endpoints for cron jobs
Before this commit, it was necessary to start a admin-site container that solely calls the cron job ("manage.py ..."), and run it as a cron job.
Because manage.py requires Django to be configured, the cron jobs required all options to be configured, meaning cron jobs deployed in e.g. Kubernetes, needed the full Django config.

With this commit, instead of having to start a container image that solely calls the manage.py, HTTP endpoints are now exposed in the admin-site container, removing the need for cron-specific admin-site containers.

The compose file still uses the setup with cron-containers, as that setup does not lend itself to the new solution.