Create default site with superuser
When running the site for the first time, the superuser creation script now also creates a default site for the superuser.

Because the superuser cannot create new sites until it is assigned a site, this change makes it easier to deploy a new admin site installation - it is no longer necessary to create the site manually through the django administration interface.