Update 'Images' button to link to ISO releases
The 'Images' button in the user GUI now provides links for releases of PC and kiosk image ISOs. Previously, ISO images were uploaded to the running admin-site container and displayed in the GUI.

With this change, only URLs to the releases are configured and supplied during startup, streamlining the process.